+++ src/components/duelling-picklist/duelling-picklist.component.tsx
@@ -76,6 +76,7 @@ export const DuellingPicklist = ({
     <StyledDuellingPicklistOverlay
       disabled={disabled}
       data-component="duelling-picklist"
+      data-role="duelling-picklist-overlay"
       {...filterStyledSystemMarginProps(rest)}
     >
       {shouldDisplayLabels && (
@@ -93,7 +94,7 @@ export const DuellingPicklist = ({
           <StyledControl data-element="picklist-left-control">
             {leftControls}
           </StyledControl>
-          <StyledControl data-element="picklist-right-label">
+          <StyledControl data-element="picklist-right-control">
             {rightControls}
           </StyledControl>
         </StyledControlsContainer>